Extract maps API calls
