+++ app/lib/api/apiWrapper.ts
@@ -1,0 +1,7 @@
+import * as maps from './requests/maps';
+
+const api = {
+    maps,
+};
+
+export default api;
